+++ get_next_line.h
@@ -38,7 +38,7 @@ t_block	*init_block(void);
 int		get_char_pos(const char *str, int c);
 int		get_end_of_line(const char *str);
 void	content_move(t_block **block);
-int		end_of_line(const char *str, const int size);
+ssize_t	end_of_line(const char *str, const ssize_t size);
 void	free_all(t_block **text_block);
 
 // GNL
